Update unit tests and implementations
