+++ src/Riot/DTO/Lol/MatchlistDTO.php
@@ -9,12 +9,6 @@ use Riot\DTO\DTOInterface;
 
 final class MatchlistDTO implements DTOInterface
 {
-    private int $startIndex;
-
-    private int $totalGames;
-
-    private int $endIndex;
-
     /** @var MatchReferenceDTOCollection<MatchReferenceDTO> */
     private MatchReferenceDTOCollection $matches;
 
@@ -22,30 +16,9 @@ final class MatchlistDTO implements DTOInterface
      * @param MatchReferenceDTOCollection<MatchReferenceDTO> $matches
      */
     public function __construct(
-        int $startIndex,
-        int $totalGames,
-        int $endIndex,
         MatchReferenceDTOCollection $matches
     ) {
-        $this->startIndex = $startIndex;
-        $this->totalGames = $totalGames;
-        $this->endIndex = $endIndex;
         $this->matches = $matches;
-    }
-
-    public function getStartIndex(): int
-    {
-        return $this->startIndex;
-    }
-
-    public function getTotalGames(): int
-    {
-        return $this->totalGames;
-    }
-
-    public function getEndIndex(): int
-    {
-        return $this->endIndex;
     }
 
     /**
@@ -59,10 +32,7 @@ final class MatchlistDTO implements DTOInterface
     public static function createFromArray(array $data): self
     {
         return new self(
-            $data['startIndex'],
-            $data['totalGames'],
-            $data['endIndex'],
-            MatchReferenceDTOCollection::createFromArray($data['matches']),
+            MatchReferenceDTOCollection::createFromArray($data),
         );
     }
 }
