+++ tests/Unit/DTO/Lol/MatchlistDTOTest.php
@@ -12,16 +12,8 @@ final class MatchlistDTOTest extends TestCase
 {
     public function testCreateFromArrayCreatesProperObject(): void
     {
-        $data = [
-            'matches' => [],
-            'startIndex' => 0,
-            'endIndex' => 100,
-            'totalGames' => 169,
-        ];
+        $data = [];
         $object = MatchlistDTO::createFromArray($data);
         self::assertInstanceOf(MatchReferenceDTOCollection::class, $object->getMatches());
-        self::assertSame(0, $object->getStartIndex());
-        self::assertSame(100, $object->getEndIndex());
-        self::assertSame(169, $object->getTotalGames());
     }
 }
